Split TabList and Tab snapshot tests
