+++ packages/experimental/TabList/src/Tab/useTab.ts
@@ -100,9 +100,9 @@ export const useTab = (props: TabProps): TabInfo => {
       accessible: accessible ?? true,
       accessibilityRole: 'tab',
       accessibilityActions: accessibilityActionsProp,
-      accessibilityPositionInSet: accessibilityPositionInSet ?? tabKeys.findIndex((key) => key === tabKey) + 1,
+      accessibilityPositionInSet: accessibilityPositionInSet ?? tabKeys?.findIndex((key) => key === tabKey) + 1,
       accessibilityState: getAccessibilityState(isDisabled, selectedKey === tabKey, accessibilityState),
-      accessibilitySetSize: accessibilitySetSize ?? tabKeys.length,
+      accessibilitySetSize: accessibilitySetSize ?? tabKeys?.length,
       disabled: isDisabled,
       focusable: !isDisabled ?? true,
       icon: icon,
